Array typing improvement

+++ mava/networks.py
@@ -193,7 +193,7 @@ class DiscreteActionEpsGreedyMaskedHead(nn.Module):
         # get num avail actions to generate probabilities for choosing
         n_available_actions = jnp.sum(masked_uniform_action_probs, axis=-1)[..., jnp.newaxis]
         # divide with sum along axis to get uniform per action choice
-        masked_uniform_action_probs = masked_uniform_action_probs.astype("int32") / inner_sum
+        masked_uniform_action_probs = masked_uniform_action_probs.astype(int) / inner_sum
 
         # GREEDY PART (1-eps %)
         # set masked actions to value not chosen by argmax
